tweak for cross-platform
