+++ src/lgfx/lgfx_font_support.hpp
@@ -21,6 +21,9 @@ Contributors:
 #define LGFX_FONT_SUPPORT_HPP_
 
 #include "../Fonts/lgfx_fonts.hpp"
+#include <string>
+#include <cmath>
+#include <stdarg.h>
 
 namespace lgfx
 {
@@ -43,6 +46,177 @@ namespace lgfx
   };
 
 //----------------------------------------------------------------------------
+
+  struct VLWfont : public RunTimeFont {
+    uint16_t gCount;     // Total number of characters
+    uint16_t yAdvance;   // Line advance
+    uint16_t spaceWidth; // Width of a space character
+    int16_t  ascent;     // Height of top of 'd' above baseline, other characters may be taller
+    int16_t  descent;    // Offset to bottom of 'p', other characters may have a larger descent
+    uint16_t maxAscent;  // Maximum ascent found in font
+    uint16_t maxDescent; // Maximum descent found in font
+
+    // These are for the metrics for each individual glyph (so we don't need to seek this in file and waste time)
+    uint16_t* gUnicode  = nullptr;  //UTF-16 code, the codes are searched so do not need to be sequential
+    uint8_t*  gWidth    = nullptr;  //cwidth
+    uint8_t*  gxAdvance = nullptr;  //setWidth
+    int8_t*   gdX       = nullptr;  //leftExtent
+    uint32_t* gBitmap   = nullptr;  //file pointer to greyscale bitmap
+
+    DataWrapper* _fontData = nullptr;
+    bool _fontLoaded = false; // Flags when a anti-aliased font is loaded
+
+    font_type_t getType(void) const override { return ft_vlw;  } 
+
+    void getDefaultMetric(FontMetrics *metrics) const override {
+      metrics->x_offset  = 0;
+      metrics->y_offset  = 0;
+      metrics->baseline  = maxAscent;
+      metrics->y_advance = yAdvance;
+      metrics->height    = yAdvance;
+    }
+
+    virtual ~VLWfont() {
+      unloadFont();
+    }
+
+    bool unloadFont(void) override {
+      _fontLoaded = false;
+      if (gUnicode)  { heap_free(gUnicode);  gUnicode  = nullptr; }
+      if (gWidth)    { heap_free(gWidth);    gWidth    = nullptr; }
+      if (gxAdvance) { heap_free(gxAdvance); gxAdvance = nullptr; }
+      if (gdX)       { heap_free(gdX);       gdX       = nullptr; }
+      if (gBitmap)   { heap_free(gBitmap);   gBitmap   = nullptr; }
+      if (_fontData) {
+        _fontData->preRead();
+        _fontData->close();
+        _fontData->postRead();
+        _fontData = nullptr;
+      }
+      return true;
+    }
+
+    bool getUnicodeIndex(uint16_t unicode, uint16_t *index) const
+    {
+      auto poi = std::lower_bound(gUnicode, &gUnicode[gCount], unicode);
+      *index = std::distance(gUnicode, poi);
+      return (*poi == unicode);
+    }
+
+    bool updateFontMetric(FontMetrics *metrics, uint16_t uniCode) const override {
+      uint16_t gNum = 0;
+      if (getUnicodeIndex(uniCode, &gNum)) {
+        if (gWidth && gxAdvance && gdX[gNum]) {
+          metrics->width     = gWidth[gNum];
+          metrics->x_advance = gxAdvance[gNum];
+          metrics->x_offset  = gdX[gNum];
+        } else {
+          auto file = _fontData;
+
+          file->preRead();
+
+          file->seek(28 + gNum * 28);  // headerPtr
+          uint32_t buffer[6];
+          file->read((uint8_t*)buffer, 24);
+          metrics->width    = __builtin_bswap32(buffer[1]); // Width of glyph
+          metrics->x_advance = __builtin_bswap32(buffer[2]); // xAdvance - to move x cursor
+          metrics->x_offset  = (int32_t)((int8_t)__builtin_bswap32(buffer[4])); // x delta from cursor
+
+          file->postRead();
+        }
+        return true;
+      }
+      return false;
+    }
+
+
+    bool loadFont(DataWrapper* data) {
+      _fontData = data;
+      {
+        uint32_t buf[6];
+        data->read((uint8_t*)buf, 6 * 4); // 24 Byte read
+
+        gCount   = __builtin_bswap32(buf[0]); // glyph count in file
+                 //__builtin_bswap32(buf[1]); // vlw encoder version - discard
+        yAdvance = __builtin_bswap32(buf[2]); // Font size in points, not pixels
+                 //__builtin_bswap32(buf[3]); // discard
+        ascent   = __builtin_bswap32(buf[4]); // top of "d"
+        descent  = __builtin_bswap32(buf[5]); // bottom of "p"
+      }
+
+      // These next gFont values might be updated when the Metrics are fetched
+      maxAscent  = ascent;   // Determined from metrics
+      maxDescent = descent;  // Determined from metrics
+      yAdvance   = std::max((int)yAdvance, ascent + descent);
+      spaceWidth = yAdvance * 2 / 7;  // Guess at space width
+
+//ESP_LOGI("LGFX", "ascent:%d  descent:%d", gFont.ascent, gFont.descent);
+
+      if (!gCount) return false;
+
+//ESP_LOGI("LGFX", "font count:%d", gCount);
+
+      uint32_t bitmapPtr = 24 + (uint32_t)gCount * 28;
+
+      gBitmap   = (uint32_t*)heap_alloc_psram( gCount * 4); // seek pointer to glyph bitmap in the file
+      gUnicode  = (uint16_t*)heap_alloc_psram( gCount * 2); // Unicode 16 bit Basic Multilingual Plane (0-FFFF)
+      gWidth    =  (uint8_t*)heap_alloc_psram( gCount );    // Width of glyph
+      gxAdvance =  (uint8_t*)heap_alloc_psram( gCount );    // xAdvance - to move x cursor
+      gdX       =   (int8_t*)heap_alloc_psram( gCount );    // offset for bitmap left edge relative to cursor X
+
+      if (gBitmap  ) gBitmap   = (uint32_t*)heap_alloc( gCount * 4); // seek pointer to glyph bitmap in the file
+      if (gUnicode ) gUnicode  = (uint16_t*)heap_alloc( gCount * 2); // Unicode 16 bit Basic Multilingual Plane (0-FFFF)
+      if (gWidth   ) gWidth    =  (uint8_t*)heap_alloc( gCount );    // Width of glyph
+      if (gxAdvance) gxAdvance =  (uint8_t*)heap_alloc( gCount );    // xAdvance - to move x cursor
+      if (gdX      ) gdX       =   (int8_t*)heap_alloc( gCount );    // offset for bitmap left edge relative to cursor X
+
+      if (!gUnicode
+       || !gBitmap
+       || !gWidth
+       || !gxAdvance
+       || !gdX) {
+//ESP_LOGE("LGFX", "can not alloc font table");
+        return false;
+      }
+
+      _fontLoaded = true;
+
+      size_t gNum = 0;
+      _fontData->seek(24);  // headerPtr
+      uint32_t buffer[7];
+      do {
+        _fontData->read((uint8_t*)buffer, 7 * 4); // 28 Byte read
+        uint16_t unicode = __builtin_bswap32(buffer[0]); // Unicode code point value
+        uint32_t w = (uint8_t)__builtin_bswap32(buffer[2]); // Width of glyph
+        if (gUnicode)   gUnicode[gNum]  = unicode;
+        if (gWidth)     gWidth[gNum]    = w;
+        if (gxAdvance)  gxAdvance[gNum] = (uint8_t)__builtin_bswap32(buffer[3]); // xAdvance - to move x cursor
+        if (gdX)        gdX[gNum]       =  (int8_t)__builtin_bswap32(buffer[5]); // x delta from cursor
+
+        uint16_t height = __builtin_bswap32(buffer[1]); // Height of glyph
+        if ((unicode > 0xFF) || ((unicode > 0x20) && (unicode < 0xA0) && (unicode != 0x7F))) {
+          int16_t dY =  (int16_t)__builtin_bswap32(buffer[4]); // y delta from baseline
+//ESP_LOGI("LGFX", "unicode:%x  dY:%d", unicode, dY);
+          if (maxAscent < dY) {
+            maxAscent = dY;
+          }
+          if (maxDescent < (height - dY)) {
+//ESP_LOGI("LGFX", "maxDescent:%d", maxDescent);
+            maxDescent = height - dY;
+          }
+        }
+
+        if (gBitmap)  gBitmap[gNum] = bitmapPtr;
+        bitmapPtr += w * height;
+      } while (++gNum < gCount);
+
+      yAdvance = maxAscent + maxDescent;
+
+//ESP_LOGI("LGFX", "maxDescent:%d", maxDescent);
+      return true;
+    }
+  };
+
 //----------------------------------------------------------------------------
 
   template <class Base>
@@ -103,7 +277,7 @@ namespace lgfx
 
         if (!_font->updateFontMetric(&_font_metrics, uniCode)) continue;
         if (left == 0 && right == 0 && _font_metrics.x_offset < 0) left = right = -_font_metrics.x_offset;
-        right = left + std::max((int32_t)_font_metrics.x_advance, _font_metrics.width + _font_metrics.x_offset);
+        right = left + std::max<int>(_font_metrics.x_advance, _font_metrics.width + _font_metrics.x_offset);
         left += _font_metrics.x_advance;
       } while (*(++string));
 
@@ -163,10 +337,10 @@ namespace lgfx
       _filled_x = 0;
       switch (fontdata[font]->getType()) {
       default:
-      case font_type_t::ft_glcd: return drawCharGLCD(this, x, y, uniCode, &_text_style, fontdata[font]);
-      case font_type_t::ft_bmp:  return drawCharBMP( this, x, y, uniCode, &_text_style, fontdata[font]);
-      case font_type_t::ft_rle:  return drawCharRLE( this, x, y, uniCode, &_text_style, fontdata[font]);
-      case font_type_t::ft_bdf:  return drawCharBDF( this, x, y, uniCode, &_text_style, fontdata[font]);
+      case IFont::font_type_t::ft_glcd: return drawCharGLCD(this, x, y, uniCode, &_text_style, fontdata[font]);
+      case IFont::font_type_t::ft_bmp:  return drawCharBMP( this, x, y, uniCode, &_text_style, fontdata[font]);
+      case IFont::font_type_t::ft_rle:  return drawCharRLE( this, x, y, uniCode, &_text_style, fontdata[font]);
+      case IFont::font_type_t::ft_bdf:  return drawCharBDF( this, x, y, uniCode, &_text_style, fontdata[font]);
       }
     }
 
@@ -200,7 +374,7 @@ namespace lgfx
 
     [[deprecated("use setFont(&fonts::Font0)")]]
     void setTextFont(int f) {
-      if (f == 1 && _font && _font->getType() == ft_gfx) return;
+      if (f == 1 && _font && _font->getType() == IFont::font_type_t::ft_gfx) return;
 
       setFont(fontdata[f]);
     }
@@ -235,11 +409,11 @@ namespace lgfx
 
       switch (font->getType()) {
       default:
-      case font_type_t::ft_glcd: fpDrawChar = drawCharGLCD;  break;
-      case font_type_t::ft_bmp:  fpDrawChar = drawCharBMP;   break;
-      case font_type_t::ft_rle:  fpDrawChar = drawCharRLE;   break;
-      case font_type_t::ft_bdf:  fpDrawChar = drawCharBDF;   break;
-      case font_type_t::ft_gfx:  fpDrawChar = drawCharGFXFF; break;
+      case IFont::font_type_t::ft_glcd: fpDrawChar = drawCharGLCD;  break;
+      case IFont::font_type_t::ft_bmp:  fpDrawChar = drawCharBMP;   break;
+      case IFont::font_type_t::ft_rle:  fpDrawChar = drawCharRLE;   break;
+      case IFont::font_type_t::ft_bdf:  fpDrawChar = drawCharBDF;   break;
+      case IFont::font_type_t::ft_gfx:  fpDrawChar = drawCharGFXFF; break;
       }
     }
 
@@ -563,7 +737,7 @@ namespace lgfx
     FontMetrics _font_metrics = { 6, 6, 0, 8, 8, 0, 7 }; // Font0 Metric
     const IFont* _font = &fonts::Font0;
 
-    IFont* _dynamic_font = nullptr;  // run-time generated font
+    RunTimeFont* _dynamic_font = nullptr;  // run-time generated font
     FileWrapper _font_file;
     PointerWrapper _font_data;
 
@@ -1016,8 +1190,8 @@ namespace lgfx
       int32_t left  = 0;
       int32_t right = 0;
       if (fillbg) {
-        left  = std::max(me->_filled_x, x + (xoffset < 0 ? xoffset : 0));
-        right = x + std::max((int32_t)(w * style->size_x + xoffset), (int32_t)(xAdvance));
+        left  = std::max<int>(me->_filled_x, x + (xoffset < 0 ? xoffset : 0));
+        right = x + std::max<int>(w * style->size_x + xoffset, xAdvance);
       }
 
       x += xoffset;
@@ -1124,7 +1298,7 @@ namespace lgfx
       int32_t right = 0;
       if (fillbg) {
         left  = std::max(me->_filled_x, x + (xoffset < 0 ? xoffset : 0));
-        right = x + std::max((int32_t)(w * style->size_x + xoffset), (int32_t)(xAdvance));
+        right = x + std::max<int>(w * style->size_x + xoffset, xAdvance);
       }
       me->_filled_x = right;
