+++ src/lgfx/platforms/lgfx_spi_esp32.hpp
@@ -28,6 +28,7 @@ Contributors:
 #if defined (ARDUINO) // Arduino ESP32
  #include <SPI.h>
 #else
+ #include <esp_log.h>
  #include <driver/spi_master.h>
  #if ESP_IDF_VERSION_MAJOR > 3
   #include <driver/spi_common_internal.h>
